feat: move ServiceException to main file library
