+++ lib/api_request_helper_flutter.dart
@@ -1,2 +1,1 @@
 export 'src/api_request_helper_flutter.dart';
-export 'src/service_exception.dart';
